requirements.txt
# Production dependencies
requests>=2.28.0
openpyxl>=3.1.0
flask>=3.0.0

# Cascade pipeline dependencies (optional - for --cascade mode)
python-dotenv>=1.0.0
google-search-results>=2.4.0  # SerpApi for Google Hotels
apify-client>=1.6.0           # Apify for Booking.com scraping
